+++ src/views/InvolvementProfile/index.jsx
@@ -18,7 +18,7 @@ import { useEffect, useRef, useState } from 'react';
 import Cropper from 'react-cropper';
 import Dropzone from 'react-dropzone';
 import { useParams } from 'react-router';
-import involvementService from 'services/activity';
+import involvementService from 'services/involvements';
 import { AuthGroup } from 'services/auth';
 import emailsService from 'services/emails';
 import membershipService from 'services/membership';
